@@ -1,3 +1,6 @@
+# typed: false
+# frozen_string_literal: true
+
 module IPWhitelist
   module Controller
     def check_ip_whitelist
@@ -5,18 +8,18 @@ module IPWhitelist
         # Since requests can come through cloudflare, try that IP first.
         ip = IPAddr.new(request.headers["CF-Connecting-IP"] || request.remote_ip)
 
-        return if current_user.ip_whitelist.detect{ |ip_or_range|
+        return if current_user.ip_whitelist.detect  do |ip_or_range|
           # Check if the IP is equal to the whitelisted IP, or is within the
           # whitelisted IP range.
           (ip_or_range.is_a?(IPAddr) ? ip_or_range : IPAddr.new(ip_or_range)) === ip
-        }
+        end
 
-        Rails.logger.info("#{current_user.username}'s IP (#{ip}) is not in authorized list (#{current_user.ip_whitelist.join(", " )})")
+        Rails.logger.info("#{current_user.username}'s IP (#{ip}) is not in authorized list (#{current_user.ip_whitelist.join(", ")})")
 
-        if redirect = IPWhitelist.redirect
+        if (redirect = IPWhitelist.redirect)
           redirect_to redirect
         else
-          render text: "Not Authorized", status: 401
+          render plain: "Not Authorized", status: :unauthorized
         end
       end
     end
