@@ -1,3 +1,6 @@
+# typed: strong
+# frozen_string_literal: true
+
 module IPWhitelist
-  VERSION = '0.0.10'
+  VERSION = "1.0.0"
 end
